fixed all pylint and pep8 errors except for the two no-member errors returned in __init__.py which I can't figure out
