+++ pylmod/base.py
@@ -1,14 +1,12 @@
-#!/usr/bin/python
 """
- Python interface to Stellar Grade Book module
 
- Defines the class StellarGradeBook
+
+Defines the class StellarGradeBook
 """
 
 import json
 import logging
 import requests
-import time
 
 
 log = logging.getLogger(__name__)  # pylint: disable=C0103
@@ -84,176 +82,6 @@ class Base(object):
             log.info(msg)
             raise Exception(msg)
         return gradebook_id['data']['gradebookId']
-
-    def get_sections(self, gradebookid='', simple=False):
-        """
-        return list of sections for a given gradebook,
-        specified by a gradebookid.
-        sample return:
-        [
-          {
-            "name": "Unassigned",
-            "editable": false,
-            "members": null,
-            "shortName": "def",
-            "staffs": null,
-            "groupId": 1293925
-          }
-        ]
-        """
-        params = dict(includeMembers='false')
-
-        sdat = self.get(
-            'sections/{gradebookId}',
-            params=params,
-            gradebookId=gradebookid or self.gradebookid
-        )
-
-        if simple:
-            return [{'SectionName': x['name']} for x in sdat['data']]
-        return sdat['data']
-
-    def get_section_by_name(self, section_name):
-        """
-        return section for a given section name
-
-        :param section_name:
-        :return:
-        """
-        sections = self.get_sections()
-        for sec in sections:
-            if sec['name'] == section_name:
-                return sec['groupId'], sec
-        return None, None
-
-    def _spreadsheet2gradebook_multi(  # pylint: disable=too-many-locals
-            self, creader, email_field, non_assignment_fields
-    ):
-        """
-        Helper function: Transfer grades from spreadsheet using
-        multiGrades (multiple students at a time). We do this by
-        creating a large array containing all grades to transfer, then
-        make one call to the gradebook API.
-        """
-        assignments = self.get_assignments()
-        students = self.get_students()
-        assignment2id = {}
-        garray = []
-        for cdat in creader:
-            email = cdat[email_field]
-            sid, _ = self.get_student_by_email(email, students)
-            if sid is None:
-                log.warning(
-                    'Error in spreadsheet2gradebook: cannot find '
-                    'student id for email="%s"\n', email
-                )
-            for field in cdat.keys():
-                if field in non_assignment_fields:
-                    continue
-                if field not in assignment2id:
-                    aid, _ = self.get_assignment_by_name(
-                        field, assignments=assignments
-                    )
-                    if aid is None:
-                        name = field
-                        shortname = field[0:3] + field[-2:]
-                        log.info('calling create_assignment from multi')
-                        resp = self.create_assignment(
-                            name, shortname, 1.0, 100.0, '12-15-2013'
-                        )
-                        if (
-                                not resp.get('data', '') or
-                                'assignmentId' not in resp.get('data')
-                        ):
-                            log.warning(
-                                'Failed to create assignment %s', name
-                            )
-                            log.info(resp)
-                            msg = (
-                                "Error ! Failed to create assignment %s", name
-                            )
-                            log.critical(msg)
-                            raise Exception(msg)
-                        aid = resp['data']['assignmentId']
-                    log.info("Assignment %s has Id=%s", field, aid)
-                    assignment2id[field] = aid
-
-                aid = assignment2id[field]
-                successful = True
-                try:
-                    gradeval = float(cdat[field]) * 1.0
-                except ValueError as err:
-                    log.exception(
-                        "Failed in converting grade for student %s"
-                        ", cdat=%s, err=%s", sid, cdat, err
-                    )
-                    successful = False
-                if successful:
-                    garray.append(
-                        {"studentId": sid,
-                         "assignmentId": aid,
-                         "numericGradeValue": gradeval,
-                         "mode": 2,
-                         "isGradeApproved": False}
-                    )
-        log.info(
-            'Data read from file, doing multiGrades API '
-            'call (%d grades)', len(garray)
-        )
-        tstart = time.time()
-        resp = self.multi_grade(garray)
-        duration = time.time() - tstart
-        log.info(
-            'multiGrades API call done (%d bytes returned) '
-            'dt=%6.2f seconds.', len(json.dumps(resp)), duration
-        )
-        return resp, duration
-
-    def _spreadsheet2gradebook_slow(  # pylint: disable=too-many-locals
-            self, creader, email_field, non_assignment_fields
-    ):
-        """
-        Helper function: Transfer grades from spreadsheet one at a time
-        """
-        assignments = self.get_assignments()
-        assignment2id = {}
-        students = self.get_students()
-        for cdat in creader:
-            email = cdat[email_field]
-            sid, _ = self.get_student_by_email(email, students)
-            for field in cdat.keys():
-                if field in non_assignment_fields:
-                    continue
-                if field not in assignment2id:
-                    assignment_id, _ = self.get_assignment_by_name(
-                        field, assignments=assignments
-                    )
-                    if assignment_id is None:
-                        name = field
-                        shortname = field[0:3] + field[-2:]
-                        resp = self.create_assignment(
-                            name, shortname, 1.0, 100.0, '12-15-2013'
-                        )
-                        if 'assignmentId' not in resp['data']:
-                            log.info(resp)
-                            msg = (
-                                "Error ! Failed to create assignment %s", name
-                            )
-                            log.error(msg)
-                            raise Exception(msg)
-                        assignment_id = resp['data']['assignmentId']
-                        log.info("Assignment %s has Id=%s",
-                                 field, assignment_id
-                                )
-                    assignment2id[field] = assignment_id
-
-                assignment_id = assignment2id[field]
-                gradeval = float(cdat[field]) * 1.0
-                log.info(
-                    "--> Student %s assignment %s grade %s",
-                    email, field, gradeval
-                )
-                self.set_grade(assignment_id, sid, gradeval)
 
     def rest_action(self, func, url, **kwargs):
         """Routine to do low-level REST operation, with retry"""
